DataTool/requirements.txt
Pillow==4.0.0
tornado==4.4.2
pymysql